+++ src/main/java/com/example/homeservice/Repositories/RoomRepositories.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomRepositories extends JpaRepository<Room,Long> {
+public interface RoomRepositories extends JpaRepository<Room,Integer> {
 }
