+++ src/main/java/com/example/homeservice/Service/RoomService.java
@@ -4,7 +4,7 @@ import com.example.homeservice.Home.Room;
 import com.example.homeservice.Request.RoomRequest;
 
 public interface RoomService {
-    Room createRoom(Long homeId, RoomRequest roomRequest);
-    Room putRoom(Long roomId,RoomRequest roomRequest);
-    void deleteRoom(Long id);
+    Room createRoom(Integer homeId, RoomRequest roomRequest);
+    Room putRoom(Integer roomId,RoomRequest roomRequest);
+    void deleteRoom(Integer id);
 }
